Se agregan los test de division multipliacion y resta

+++ src/main/java/co/com/sofka/app/calculator/BasicCalculator.java
@@ -22,7 +22,7 @@ public class BasicCalculator {
         return number1 * number2;
     }
 
-    public double division(Long number1, Long number2) {
+    public double division(double number1, double number2) {
         logger.info("Summing {} / {}", number1, number2);
         try {
             return number1 / number2;
